requirements.txt
django==3.2
djangorestframework==3.12.2
djangorestframework-simplejwt==5.1.0
django-cors-headers==3.11.0
djangorestframework-api-key==2.1.0